+++ node/index.ts
@@ -8,6 +8,7 @@ import { Service, method, LRUCache } from '@vtex/api'
 
 import { Clients } from './clients'
 import { errorHandler } from './middlewares/errorHandler'
+import { setupLogger } from './middlewares/setupLogger'
 import { mutations, queries, resolvers } from './resolvers'
 import { schemaDirectives } from './directives'
 import { auth } from './middlewares/auth'
@@ -48,7 +49,9 @@ const clients: ClientsConfig<Clients> = {
 }
 
 declare global {
-  type Context = ServiceContext<Clients, State>
+  type Context = ServiceContext<Clients, State> & {
+    logger?: any // Dynatrace logger instance from setupLogger middleware
+  }
 
   interface State extends RecorderState {
     // Added in the state via graphql directive or auth middleware when request has vtexidclientautcookie
@@ -62,31 +65,31 @@ export default new Service<Clients, State, ParamsContext>({
   clients,
   routes: {
     returnRequests: method({
-      POST: [errorHandler, auth, createReturn],
-      GET: [errorHandler, auth, getRequestList],
+      POST: [setupLogger, errorHandler, auth, createReturn],
+      GET: [setupLogger, errorHandler, auth, getRequestList],
     }),
     returnRequest: method({
-      GET: [errorHandler, auth, getRequest],
-      PUT: [errorHandler, auth, updateRequestStatus],
+      GET: [setupLogger, errorHandler, auth, getRequest],
+      PUT: [setupLogger, errorHandler, auth, updateRequestStatus],
     }),
     returnRequestAdditionalInfo: method({
-      GET: [errorHandler, auth, getRequestAdditionalInfo],
-      PUT: [errorHandler, auth, updateRequestAdditionalInfo],
+      GET: [setupLogger, errorHandler, auth, getRequestAdditionalInfo],
+      PUT: [setupLogger, errorHandler, auth, updateRequestAdditionalInfo],
     }),
     adjustmentNotes: method({
-      GET: [errorHandler, auth, getAdjustmentList],
-      POST: [errorHandler, auth, createAdjustment],
+      GET: [setupLogger, errorHandler, auth, getAdjustmentList],
+      POST: [setupLogger, errorHandler, auth, createAdjustment],
     }),
     adjustmentNote: method({
-      GET: [errorHandler, auth, getAdjustment],
-      PUT: [errorHandler, auth, updateAdjustmentStatus],
+      GET: [setupLogger, errorHandler, auth, getAdjustment],
+      PUT: [setupLogger, errorHandler, auth, updateAdjustmentStatus],
     }),
     adjustmentNoteAdditionalInfo: method({
-      GET: [errorHandler, auth, getAdjustmentAdditionalInfo],
-      PUT: [errorHandler, auth, updateAdjustmentAdditionalInfo],
-    }),  
+      GET: [setupLogger, errorHandler, auth, getAdjustmentAdditionalInfo],
+      PUT: [setupLogger, errorHandler, auth, updateAdjustmentAdditionalInfo],
+    }),
     keepAlive: method({
-      GET: [keepAlive],
+      GET: [setupLogger, keepAlive],
     }),
   },
   graphql: {