+++ node/utils/validatePaymentMethod.ts
@@ -18,7 +18,7 @@ export const validatePaymentMethod = (
   // When admin doesn't allow selection, PM request has to be sameAsPurchase
   if (
     !enablePaymentMethodSelection &&
-    refundPaymentMethod !== 'sameAsPurchase' && 
+    refundPaymentMethod !== 'sameAsPurchase' &&
     refundPaymentMethod !== 'giftCard'
   ) {
     throw new ResolverError(