+++ react/common/messages.ts
@@ -3,17 +3,17 @@ import { defineMessages } from 'react-intl'
 export const navigationMessages = defineMessages({
   adminNavigationLabel: {
     id: 'admin/return-app.navigation.label',
-    defaultMessage: 'Returns'
+    defaultMessage: 'Returns',
   },
   navigationLabelRequests: {
     id: 'navigation.labelRequests',
-    defaultMessage: 'Requests'
-  }
+    defaultMessage: 'Requests',
+  },
 })
 
 export const verificationStatusMessages = defineMessages({
   cancelled: {
     id: 'return-app.return-request-details.table.verification-status.cancelled',
-    defaultMessage: 'Cancelled'
-  }
-}) 
+    defaultMessage: 'Cancelled',
+  },
+})